LRCI-635 Extract stable-job specific logic to a new TopLevelBuild class, PortalPullRequestTesterTopLevelBuild
