+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BuildFactory.java
@@ -75,6 +75,9 @@ public class BuildFactory {
 				return new StandaloneTopLevelBuild(
 					url, (TopLevelBuild)parentBuild);
 			}
+
+			return new PortalPullRequestTesterTopLevelBuild(
+				url, (TopLevelBuild)parentBuild);
 		}
 
 		if ((parentBuild != null) &&
